refactored word/pdf exports
